+++ src/collision.h
@@ -8,6 +8,6 @@
 #include "direction.h"
 
 bool sprite_sprite_collision(uint8_t, uint8_t, uint8_t, uint8_t, uint8_t, uint8_t, uint8_t, uint8_t);
-bool sprite_tile_collision(uint16_t, uint16_t, Direction);
+bool sprite_tile_collision(uint8_t, uint8_t, Direction);
 
 #endif
